Added feedback and update appointment service

+++ src/main/java/com/java/medrecord/dao/AppointmentRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.java.medrecord.entity.Appointment;
 
-public interface AppointmentRepository extends JpaRepository<Appointment, Integer>{
+public interface AppointmentRepository extends JpaRepository<Appointment, Long>{
 
 }
